pyproject.toml
[project]
name = "knowlet"
version = "0.0.1"
description = "A personal knowledge base that organizes itself."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Jolan" }]
dependencies = [
    "openai>=1.40",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "typer>=0.12",
    "rich>=13.7",
    "httpx>=0.27",
    "python-frontmatter>=1.1",
    "mistune>=3.0",
    "watchdog>=4.0",
    "python-ulid>=2.7",
    "sqlite-vec>=0.1.6",
    "tomlkit>=0.13",
    "numpy>=1.26",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
]

[project.optional-dependencies]
embed = [
    "sentence-transformers>=2.7",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
]

[project.scripts]
knowlet = "knowlet.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["knowlet"]

[tool.hatch.build.targets.sdist]
include = ["knowlet", "README.md", "LICENSE"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]
